hw1-styled-components-divied
