--- homeworks/week22/hw1/.eslintrc.json
@@ -1,8 +1,0 @@
-{
-  "extends": [
-      "react-app"
-   ],
-   "rules": {
-     "react/prop-types": "warn"
-   }
- }
