+++ homeworks/week22/hw1/src/components/Footer/Footer.js
@@ -1,17 +1,5 @@
 import React from 'react';
-import styled from 'styled-components';
-
-const FooterContainer = styled.div`
-  width: 100vw;
-  padding: 10px 0;
-  background: #361058;
-  color: #e7edf3;
-  font-size: 12px;
-  font-family: 'Comfortaa', sans-serif;
-  text-align: center;
-  letter-spacing: 1px;
-  margin-top: 50px;
-`;
+import { FooterContainer } from './FooterStyle';
 
 const Footer = () => {
   return (
