+++ homeworks/week22/hw1/src/components/Footer/FooterStyle.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const FooterContainer = styled.div`
+  width: 100vw;
+  padding: 10px 0;
+  background: #361058;
+  color: #e7edf3;
+  font-size: 12px;
+  font-family: 'Comfortaa', sans-serif;
+  text-align: center;
+  letter-spacing: 1px;
+  margin-top: 50px;
+`;
