--- homeworks/week22/hw1/src/index.css
@@ -1,17 +1,0 @@
-html,
-body {
-  margin: 0;
-  padding: 0;
-  height: 100%;
-  width: 100vw;
-  -webkit-box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  box-sizing: border-box;
-}
-#root {
-  height: 100%;
-  width: 100%;
-}
-.content {
-  padding-bottom: 20px;
-}
