+++ homeworks/week22/hw1/src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import App from './components/App/App';
 import { ThemeProvider } from "styled-components";
