+++ homeworks/week22/hw1/src/pages/RegisterPage/index.js
@@ -1,1 +1,1 @@
-export { RegisterPage } from './RegisterPage';
+export { default } from './RegisterPage';
